flask-server/requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.8
MarkupSafe==2.1.5
nibabel==5.2.1
numpy==2.0.1
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.10
pymongo==4.10.1
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
scipy==1.13.0
setuptools==75.6.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.0.3
wheel==0.45.1
